pyproject.toml
[build-system]
requires = [
	"setuptools>=61.0",
	"setuptools_scm[toml]>=3.4.3",
	"wheel",
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_scheme = "release-branch-semver"
local_scheme = "no-local-version"

[project]
name = "gw-catalog-plotter"
dynamic = ["version"]
authors = [
	{ name = "Daniel Williams", email = "daniel.williams@ligo.org" },
	]
description = "Utilities for plotting gravitational wave event catalogues."
readme = "README.md"
license = "MIT" 
classifiers = [
	"Programming Language :: Python :: 3",
	"Operating System :: OS Independent",
        'Environment :: Console',
        'Intended Audience :: Science/Research',
        'Topic :: Scientific/Engineering :: Astronomy',
        'Topic :: Scientific/Engineering :: Information Analysis',
        'Natural Language :: English',
]
requires-python = ">=3.9"
dependencies = [
	     "click",
	     "matplotlib",
		 "pyyaml",
		 "h5py",
		 "pyfonts",
		 "scipy",
		 "pesummary",
]

[project.optional-dependencies]
docs = [
	"sphinx",
	"kentigern >= 0.3.0",
        "numpydoc",
	"sphinx-click",
	"sphinx-multiversion"
]

[tool.ruff]
line-length = 88
target-version = "py39"
exclude = ["build", "dist", ".venv", "environment", "docs/_build", "__pycache__"]
extend-ignore = ["E203"]
extend-select = ["B", "C", "E", "F", "I", "UP"]

[tool.setuptools.packages.find]
where = ["python"]

[project.scripts]
catplot = "catplot.cli:catplot"

[project.urls]
"Source code" = "https://github.com/transientlunatic/catalog-plotter"